requirements.txt
black
openai
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
# For AWS Bedrock integration (when implemented)
# boto3==1.28.57
# botocore==1.31.57
pygame==2.1.2
boto3
